% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_course.R
\name{install_course}
\alias{install_course}
\title{Install course materials}
\usage{
install_course(path = ".")
}
\arguments{
\item{path}{The path on your computer where you would like the course
installed}
}
\description{
`install_course()` will install the course materials for the course Mastering
R for Epidemiologic Research on your computer. Then, it will open a new
RStudio Project containing the files you'll need.
}
